requirements.txt
Django==1.8.4
argparse==1.2.1
django-crispy-forms==1.5.2
psycopg2==2.6.1
wsgiref==0.1.2
